% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFNC.R
\name{AFNC}
\alias{AFNC}
\title{Performs AFNC}
\usage{
AFNC(p.value, alpha = 0.05, beta = 0.1, cd, c0 = NULL)
}
\arguments{
\item{p.value}{d-vector of p-values.}

\item{alpha}{level of family-wise error rate for false positive control using the Bonferroni.}

\item{beta}{level of signal missing rate for false negative control using the AFNC.}

\item{cd}{\eqn{c_d} for controlling Type I error rate under the global null, must be pre-computed using {estimate.cd}.}

\item{c0}{maximum number of the most significant p-values considered.  If {NULL}, {c0} will be set to \eqn{\min( \max(5000,0.1*d), d )}, such that at least 10\% of the p-values are considered.}
}
\value{
\item{bonferroni}{Two vectors ({index}, {p.value}).  {index} is a vector of indices of variables (SNPs) selected by the Bonferroni, and {p.value} is the corresponding p-values.}
 \item{afnc}{Two vectors ({index}, {p.value}).  {index} is a vector of indices of variables (SNPs) selected by the AFNC, and {p.value} is the corresponding p-values.}
 \item{t.alpha}{Two values ({rank}, {p.value}).  {rank} specifies the rank by which all variants ranked at or before \eqn{t_\alpha} are retained for controlling false positives by Bonferroni.  {p.value} is the corresponding p-value, such that all variants less than or equal to {p.value} are retained.}
 \item{T.fn}{Two values ({rank}, {p.value}). {rank} specifies the rank by which all variants ranked at or before \eqn{t_\alpha} are retained for adaptive false negative control (AFNC).  {p.value} is the corresponding p-value, such that all variants less than or equal to {p.value} are retained.}
 \item{signal.proportion}{The estimated signal proportion \eqn{\hat{\pi}}.}
 \item{number.signals}{The estimated number of signals \eqn{\hat{s} = \hat{\pi} * d}.}
}
\description{
AFNC takes a vector of p-values and determines the variables (SNPs) selected by AFNC.  The AFNC can retain a high proportion of signals by discarding variants from the Noise region using adaptive false negative control (Jeng et al. 2016). Proportion of signals is estimated from the data.
}
\details{
The algorithm implemented in this function is as follows.  (See Jeng et al. (2016) for further details.)
\enumerate{
  \item The p-values are ordered at decreasing significance.
  \item The signal proportion estimator \eqn{\hat{\pi}} and estimated number of signals \eqn{\hat{s} = \hat{\pi} * d} are obtained using {estimate.signal.proportion}.
  \item Two cutoff positions, \eqn{t_\alpha} and \eqn{T_fn}, are determined to separate the Signal, Indistinguishable, and Noise regions. (See Figure 1 of Jeng et al. (2016) for illustration of the Signal, Indistinguishable, and Noise regions of inference.)
  \item Finally, variables (SNPs) with ordered p-values ranked at or before \eqn{t_\alpha} are selected by Bonferroni for family-wise false positive control.  Variables (SNPs) with ordered p-values ranked at or before \eqn{T_fn} are selected by the AFNC procedure for adaptive false negative control.
}
}
\examples{
# Estimate signal proportions
set.seed(1)
cd = estimate.cd(d=length(p.value), alpha=0.05)
afnc = AFNC(p.value, alpha=0.05, beta=0.1, cd=cd)$afnc
selected = afnc$index # selected variables
selected.p.value = afnc$p.value # p-values of selected variables

}
\references{
Jeng, X.J., Daye, Z.J., Lu, W., and Tzeng, J.Y. (2016) Rare Variants Association Analysis in Large-Scale Sequencing Studies at the Single Locus Level.
}

