% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFNC.R
\name{estimate.cd}
\alias{estimate.cd}
\title{Estimates \eqn{c_d} for controlling Type I error rate under the global null.}
\usage{
estimate.cd(d, M = 10000, alpha = 0.05)
}
\arguments{
\item{d}{number of variables (SNPs).}

\item{M}{number of Monte Carlo samples. The larger the \eqn{M} the more accurate the estimate.}

\item{alpha}{significance level for false positive control.}
}
\value{
Estimate of \eqn{c_d} for controlling Type I error rate under the global null, to be used as an input in {AFNC}.
}
\description{
{estimate.cd} empirically estimates \eqn{c_d} to control the Type I error rate under the global null hypothesis that no signals exist.  In this algorithm, \eqn{M} number of Monte Carlo samples, drawn under the global null hypothesis of no signals, are used to estimate \eqn{c_d}.  The estimate of \eqn{c_d} is pre-computed in {estimate.cd} and passed to {AFNC}.
}
\examples{
# Load "AFNC" library and example data.
library("AFNC")
data(example_data)

set.seed(1)  # Set seed
cd = estimate.cd(d=length(p.value), M=10000, alpha=0.05)  # estimate c_d

}

