% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFNC.R
\name{estimate.signal.proportion}
\alias{estimate.signal.proportion}
\title{Estimates signal proportions}
\usage{
estimate.signal.proportion(p.value, cd, c0 = NULL, is.sorted = FALSE)
}
\arguments{
\item{p.value}{d-vector of p-values}

\item{cd}{\eqn{c_d} for global Type I error control}

\item{c0}{maximum number of the most significant p-values considered.  If {NULL}, {c0} will be set to \eqn{\min( \max(5000,0.1*d), d )}, such that at least 10\% of the p-values are considered.}

\item{is.sorted}{if p.value is already sorted.  If {TRUE}, p.value will not be re-sorted.}
}
\value{
\item{signal.proportion}{The estimated signal proportion \eqn{\hat{\pi}}.}
 \item{number.signals}{The estimated number of signals \eqn{\hat{s} = \hat{\pi} * d}.}
}
\description{
{estimate.signal.proportion} estimates the signal proportion \eqn{\pi} using a modified estimator based on Meinshausen and Rice (2006). The estimator does not depend on statistical normality assumptions and can adapt to \eqn{\pi} small.  (See Jeng et al. (2016) for details.)
}
\examples{
# Load "AFNC" library and example data.
library("AFNC")
data(example_data)

# Estimate signal proportions
set.seed(1)
cd = estimate.cd(d=length(p.value), M=10000, alpha=0.05) # Estimate cd
obj = estimate.signal.proportion(p.value, cd=cd)
signal.proportion = obj$signal.proportion  # Estimated signal proportion
number.signals = obj$number.signals  # Estimated number of signals

}
\references{
Jeng, X.J., Daye, Z.J., Lu, W., and Tzeng, J.Y. (2016) Rare Variants Association Analysis in Large-Scale Sequencing Studies at the Single Locus Level.

 Meinshausen M. and Rice, J. (2006) Estimating the proportion of false null hypotheses among a large number of independent tested hypotheses.  {Ann. Statist.}, 34:373-393.
}

