% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFNC.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data}
\format{A data set with a vector p-values.}
\usage{
data(example_data)
}
\description{
This dataset contains 10,000 p-values.  These data are used to illustrate the functions {AFNC} and {estimate.signal.proportion}.
}
\examples{
# Load "AFNC" library and example data.
library("AFNC")
data(example_data)


}
\keyword{datasets}

